pure/configure.ac: safer CFLAGS/CXXFLAGS munging
instead of clobbering user's settings, we now just append
our overrides.

old code worked on the wrong assumption that the only
gcc / g++ option starting with "-g" is "-g" itself.
the whole thing wasn't necessary, since the last option wins:

  gcc -g -g0

is the same as

  gcc -g0

and

  gcc -Os -O3

is the same as

  gcc -O3

these claims are based on reading gcc(1) (-g), and diffing
the output of gcc -Q $opts --help=optimizers for $opts from
"-O0", "-O3", "-O0 -O3", and "-O3 -O0".

--HG--
branch : release-cflags