acinclude.m4: Use return instead of exit
On macOS with the clang compiler from Xcode 12 or later, the configure
script gets the wrong answer when checking if signal handlers must be
reinstalled when invoked because the test uses exit without including
stdlib.h where it's defined. Clang in Xcode 12 and later considers
implicit declaration of function to be an error. Fix it by using return
instead of exit.